refactor(apex-lsp-compliant-services): extract data-owner queue substrate - W-24145209
Move the data-owner tiered read/write queue (DOQueueItem, DOQueues,
dataOwnerRead/Write and their private processing loop) and the per-URI
symbol-readiness latches (ReadinessLatch, readinessLatches, armReadiness,
resolveReadiness, clearReadiness, symbolsAreCurrent) into a new
src/worker/dataOwnerQueue.ts.

The module depends only on runtimeContext (tracing hooks) and the
compliant-services document-state cache, keeping it low in the worker
dependency graph. worker.platform.shared.ts imports the substrate and
re-exports the previously-public surface unchanged.

The handler factories (dataOwnerDocHandler / requestHandler /
effectRequestHandler) stay in place for now: they depend on the
services-acquisition effects still defined in the shared module, and
will move with that substrate in a later slice.

W-24145209
